Added UI support in book config for TreeView
Added support in the book config UI to support the tree view accessibility solution. If not present when a config is loaded, a config key of include_tree_view is added. By default , the value is false. Added checkbox that will set the value to true as requested by instructor.